fix(Solid): updates Column styles
